fix: publish-website should not continue if it fails to checkout gh-pages

+++ publish-website.sh
@@ -1,18 +1,20 @@
 #!/usr/bin/env bash
 
-git checkout gh-pages
-rm -rf css
-rm -rf img
-rm -rf js
-cp -r ../falconframework.org/* .
+if git checkout gh-pages
+then
+    rm -rf css
+    rm -rf img
+    rm -rf js
+    cp -r ../falconframework.org/* .
 
-mv compressed/* ./
-rm compressed
+    mv compressed/* ./
+    rm compressed
 
-mv css/compressed/* css/
-rm css/compressed
+    mv css/compressed/* css/
+    rm css/compressed
 
-git add *
-git commit -m 'doc: Publish website'
-git push origin gh-pages
-git checkout master
+    git add *
+    git commit -m 'doc: Publish website'
+    git push origin gh-pages
+    git checkout master
+fi
